Path Planner class
To use it, construct a new object, via PathPlanner(String[][] s)
constructor, where s is the map representation. Call the Asearch( "x1 y1
0", "x2 y2") on the created object, where x1 and y1 are the coordinates
of the starting point, and x2, y2 the coordinates of the target. This is
the algorithm that will search for the best path, using the A*
technique. After this, object.reconstructPath() will return an Arraylist
of type String that contains the nodes that need to be traversed to get
to the destination.