Explored area and trash mapping
Explored area is disabled now. If I run it normally, not as a thread,
then it slows down the wall follower and causes it to bump into walls.
If I run it as a thread, it produces some random 'walls' every now and
then.  However, if you want to run it threaded, just uncomment the line
in WallFollower.run that sais t1.start().